Apache camel enable sqs propagation (#2102)
* adding AWS SQS tests to Apache Camel instrumentation

* code review changes


* SQS context propagation for Apache Camel instrumentation

* code review - removed not needed dep
