+++ instrumentation/apache-camel-2.20/javaagent-unittests/apache-camel-2.20-javaagent-unittests.gradle
@@ -1,0 +1,10 @@
+apply from: "$rootDir/gradle/java.gradle"
+
+dependencies {
+  testImplementation project(':instrumentation:apache-camel-2.20:javaagent')
+  testImplementation group: 'org.apache.camel', name: 'camel-core', version: '2.20.1'
+  testImplementation group: 'org.apache.camel', name: 'camel-aws', version: '2.20.1'
+
+  testImplementation deps.opentelemetryTraceProps
+  testImplementation group: 'org.assertj', name: 'assertj-core', version: '3.18.1'
+}
