+++ instrumentation/apache-camel-2.20/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/apachecamel/ApacheCamelInstrumentationModule.java
@@ -38,6 +38,11 @@ public class ApacheCamelInstrumentationModule extends InstrumentationModule {
     return singletonList(new CamelContextInstrumentation());
   }
 
+  @Override
+  public String[] additionalHelperClassNames() {
+    return new String[] {"io.opentelemetry.extension.trace.propagation.AwsXrayPropagator"};
+  }
+
   public static class CamelContextInstrumentation implements TypeInstrumentation {
     @Override
     public ElementMatcher<ClassLoader> classLoaderOptimization() {
