+++ settings.gradle
@@ -56,6 +56,7 @@ include ':smoke-tests'
 include ':instrumentation:akka-actor-2.5:javaagent'
 include ':instrumentation:akka-http-10.0:javaagent'
 include ':instrumentation:apache-camel-2.20:javaagent'
+include ':instrumentation:apache-camel-2.20:javaagent-unittests'
 include ':instrumentation:apache-httpasyncclient-4.1:javaagent'
 include ':instrumentation:apache-httpclient:apache-httpclient-2.0:javaagent'
 include ':instrumentation:apache-httpclient:apache-httpclient-4.0:javaagent'
